Cleaned up code while moving to compiling with avr-gcc, still using Arduino.h tho. Allow for larger rotation range by changing to larger integers

+++ OpenServo/enc.h
@@ -25,7 +25,7 @@
 
 void enc_init(void);
 /// Return the encoder position value.
-uint16_t enc_get_position_value(void);
+int32_t enc_get_position_value(void);
 
 
 #endif
